Organization of overall project by splitting up into files
