+++ Channels.h
@@ -1,9 +1,17 @@
+/*
+*  DROMatic.ino
+*  DROMatic OS Core
+*  Devin R. Olsen - Dec 31, 2016
+*  devin@devinrolsen.com
+*/
 // Channels.h
 
 #ifndef _CHANNELS_h
 #define _CHANNELS_h
 #include "Globals.h"
 #include <ArduinoJson\ArduinoJson.h> //Arduno Json (aka epic)
+
+extern int currentChannelIndex;
 
 extern JsonObject& getChannelData(JsonBuffer& b, int channelIndex = currentChannelIndex);
